requirements.txt
torch>=2.3.1
torchvision>=0.18.1
datasets>=2.21.0
matplotlib>=3.9.2
scikit-learn>=1.5.2
pypop7>=0.0.82
timm>=1.0.15
numpy==1.26.4
open-clip-torch>=3.2.0